Fix temperature reading
Looks like the 16bit temperature value is signed so that needs to be taken care of before converting
to float. Problem was that below 25 degree celsius temperatures would show as above 8000 C.

+++ src/SparkFunLSM303C.cpp
@@ -121,7 +121,7 @@ float LSM303C::readTempC()
         return NAN;
     }
 
-    temperature = (float)((valueH << 8) | valueL);
+    temperature = (float)(int16_t)((valueH << 8) | valueL);
     temperature /= 8; // 8 digits/˚C
     temperature += 25;// Reads 0 @ 25˚C
